LehuXuexi/requirements.txt
pillow==5.3.0
markdown==3.1.1
Pygments==2.2.0
selenium==3.141.0

# django
django==2.1.7
django-ckeditor==5.7.1
django-filter==2.1.0
django-guardian==2.1.0
django-crispy-forms==1.7.2

# DRF
coreapi==2.3.3
djangorestframework==3.9.4
django-cors-headers==3.0.1
djangorestframework-jwt==1.11.0